+++ internal/delivery/http/router/router.go
@@ -23,7 +23,8 @@ func New() *gin.Engine {
 	registerUserRouter(v1)
 	registerAttendanceRoutes(v1)
 	registerLeaveRequestRoutes(v1)
-	// registerPhotoRoutes(v1)
 	registerFileRoutes(v1)
+	registerPayrollSetting(v1)
+
 	return r
 }